requirements.txt
numpy==1.26.4
pandas==1.5.3
matplotlib==3.8.3
scipy==1.13.0
pyyaml==6.0.1
scikit-learn==1.4.1.post1
xgboost==2.0.3
ipykernel
neurokit2==0.2.7
statsmodels==0.14.1
PyWavelets==1.4.1